+++ libphysical/phys_gdb.h
@@ -29,9 +29,9 @@ int physical_gdb_init(struct physical_handle_data* h);
 
 int physical_gdb_finish(struct physical_handle_data* h);
 
-int physical_gdb_read(struct physical_handle_data* h, addr_t adr, void* buf, size_t len);
+int physical_gdb_read(struct physical_handle_data* h, addr_t adr, void* buf, unsigned long int len);
 
-int physical_gdb_write(struct physical_handle_data* h, addr_t adr, void* buf, size_t len);
+int physical_gdb_write(struct physical_handle_data* h, addr_t adr, void* buf, unsigned long int len);
 
 int physical_gdb_read_page(struct physical_handle_data* h, addr_t pagenum, void* buf);
