+++ libphysical/phys_ieee1394.h
@@ -29,9 +29,9 @@ int physical_ieee1394_init(struct physical_handle_data* h);
 
 int physical_ieee1394_finish(struct physical_handle_data* h);
 
-int physical_ieee1394_read(struct physical_handle_data* h, addr_t adr, void* buf, size_t len);
+int physical_ieee1394_read(struct physical_handle_data* h, addr_t adr, void* buf, unsigned long int len);
 
-int physical_ieee1394_write(struct physical_handle_data* h, addr_t adr, void* buf, size_t len);
+int physical_ieee1394_write(struct physical_handle_data* h, addr_t adr, void* buf, unsigned long int len);
 
 int physical_ieee1394_read_page(struct physical_handle_data* h, addr_t pagenum, void* buf);
